+++ lib/ui/groups/GroupPollListPanel.dart
@@ -154,7 +154,7 @@ class _GroupPollListPanelState extends State<GroupPollListPanel> implements Noti
       String? groupId = widget.group.id;
       if (StringUtils.isNotEmpty(groupId)) {
         _setGroupPollsLoading(true);
-        Groups().loadGroupPolls({groupId!}, cursor: _pollsCursor)?.then((PollsChunk? result) {
+        Polls().getGroupPolls(groupIds: {groupId!}, cursor: _pollsCursor)?.then((PollsChunk? result) {
           if (result != null) {
             if (_polls == null) {
               _polls = [];